Refactor
Refactoring of package names for clarity
