+++ app/src/main/java/com/example/appyaganaste/data/local/database/AppDatabase.kt
@@ -1,9 +1,9 @@
-package com.example.appyaganaste.data.database
+package com.example.appyaganaste.data.local.database
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
-import com.example.appyaganaste.data.database.dao.BankDao
-import com.example.appyaganaste.data.database.entity.BankEntity
+import com.example.appyaganaste.data.local.database.dao.BankDao
+import com.example.appyaganaste.data.local.database.entity.BankEntity
 
 @Database(entities = [BankEntity::class], version = 1, exportSchema = false)
 abstract class AppDatabase : RoomDatabase() {
