+++ app/src/main/java/com/example/appyaganaste/ui/bankDetail/BankDetailUI.kt
@@ -9,8 +9,6 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
@@ -18,7 +16,7 @@ import androidx.compose.ui.tooling.preview.PreviewParameterProvider
 import androidx.compose.ui.unit.dp
 import coil.compose.rememberImagePainter
 import coil.transform.CircleCropTransformation
-import com.example.appyaganaste.data.Bank
+import com.example.appyaganaste.domain.model.Bank
 
 class BankPreviewParameterProvider : PreviewParameterProvider<Bank> {
     override val values: Sequence<Bank> = sequenceOf(
